[UPDATE] 대규모 구조 변경(과도하게 구분된 디렉토리 삭제)
