+++ server/view/posts/comment.py
@@ -5,7 +5,8 @@ from flask_jwt_extended import get_jwt_identity, jwt_required
 
 from server.extensions import db
 from server.model.comment import Comment
-from server.docs.posts.comment import COMMENT_POST
+from server.model.post import Post
+from server.docs.posts import COMMENT_POST
 from server.view import unicode_safe_json_dumps
 
 
@@ -13,7 +14,7 @@ class PostComment(Resource):
 
     @swag_from(COMMENT_POST)
     @jwt_required
-    def post(self, postId):
+    def post(self, post_id):
         content = request.json['content']
 
         if content:
@@ -23,3 +24,30 @@ class PostComment(Resource):
             return unicode_safe_json_dumps({'status': '글 작성 완료.'}, 201)
         else:
             return unicode_safe_json_dumps({'status': '내용이 없습니다.'}, 400)
+
+
+class PostCommentView(Resource):
+
+    def get(self):
+        post_num = request.args.get('postId')
+        post_info = Post.quere.filter(Post.post_id == post_num).first()
+        comment = Comment.quere.filter(Comment.post_id == post_num).order_by(Post.date).all()
+
+        post_info = {
+            "content": post_info.content,
+            "user": post_info.user,
+            "image": post_info.image,
+            "date": str(post_info.date),
+            "like": post_info.like
+        }
+
+        if post_num:
+            return unicode_safe_json_dumps({
+                'post_info': post_info,
+                'comment': [
+                    {
+                        'content': comments.content,
+                        'user': comments.user,
+                        'date': comments.date
+                    } for comments in comment
+                ]}, 200)
