+++ tests/MainPage.py
@@ -1,5 +1,5 @@
 import unittest
-from server.view.mainPage.MainPage import MainPage
+from server.view.posts.mainPage import MainPage
 
 
 class MainPageTest(unittest.TestCase):
